pyproject.toml
[tool.poetry]
name = "locomote"
version = "0.1.3"
description = ""
authors = ["Donatas Rasiukevicius <donatas@rasiukevicius.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pygments = "^2.18.0"
pillow = "^10.4.0"
moviepy = "^1.0.3"
click = "^8.1.7"
typer = "^0.12.3"
gitpython = "^3.1.43"
dacite = "^1.8.1"
toml = "^0.10.2"
tiktoken = "^0.7.0"
numpy = "^2.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
locomote = "locomote.cli:app"
